+++ composeApp/src/commonMain/kotlin/dev/anthonyhfm/amethyst/devices/effects/copy/CopyChainDevice.kt
@@ -575,7 +575,7 @@ data class CopyChainDeviceState(
     val isolate: IsolationType = IsolationType.NONE,
     val timing: Timing = Timing.Rythm(Timing.Rythm.RythmTiming._1_4),
     val offsets: List<Pair<Int, Int>> = emptyList(),
-    val delayMs: Int = 0,
+    val delayMs: Long = 0,
     val gate: Float = 0.5f // 100% = 0.5f, 200% = 1.0f
 ) : DeviceState() {
     enum class CopyType {
