+++ composeApp/src/commonMain/kotlin/dev/anthonyhfm/amethyst/devices/effects/hold/HoldChainDevice.kt
@@ -204,7 +204,7 @@ class HoldChainDevice : GenericChainDevice<HoldChainDeviceState>() {
 @Serializable
 data class HoldChainDeviceState(
     val timing: Timing = Timing.Rythm(Timing.Rythm.RythmTiming._1_4),
-    val delayMs: Int = 0,
+    val delayMs: Long = 0,
     val gate: Float = 0.5f, // 100% = 0.5f, 200% = 1.0f
     val infinite: Boolean = false,
     val onRelease: Boolean = false,
